+++ udata/core/dataset/permissions.py
@@ -16,7 +16,7 @@ class OwnablePermission(Permission):
             needs.append(OrganizationAdminNeed(obj.organization.id))
             needs.append(OrganizationEditorNeed(obj.organization.id))
         elif obj.owner:
-            needs.append(UserNeed(obj.owner.id))
+            needs.append(UserNeed(obj.owner.fs_uniquifier))
 
         super(OwnablePermission, self).__init__(*needs)