+++ udata/tests/dataset/test_dataset_model.py
@@ -7,8 +7,9 @@ from mongoengine import post_save
 
 from udata.app import cache
 from udata.models import (
-    db, Dataset, License, LEGACY_FREQUENCIES, ResourceSchema
+    db, Dataset, License, LEGACY_FREQUENCIES, ResourceSchema, UPDATE_FREQUENCIES
 )
+from udata.core.dataset.models import HarvestDatasetMetadata, HarvestResourceMetadata
 from udata.core.dataset.factories import (
     ResourceFactory, DatasetFactory, CommunityResourceFactory, LicenseFactory
 )
@@ -126,10 +127,58 @@ class DatasetModelTest:
         dataset = DatasetFactory()
         assert dataset.next_update is None
 
-    def test_next_update_weekly(self):
-        dataset = DatasetFactory(frequency='weekly')
+    def test_next_update_hourly(self):
+        dataset = DatasetFactory(frequency='hourly')
+        assert_equal_dates(dataset.next_update,
+                           datetime.now() + timedelta(hours=1))
+
+    @pytest.mark.parametrize('freq', ['fourTimesADay', 'threeTimesADay', 'semidaily', 'daily'])
+    def test_next_update_daily(self, freq):
+        dataset = DatasetFactory(frequency=freq)
+        assert_equal_dates(dataset.next_update,
+                           datetime.now() + timedelta(days=1))
+
+    @pytest.mark.parametrize('freq', ['fourTimesAWeek', 'threeTimesAWeek', 'semiweekly', 'weekly'])
+    def test_next_update_weekly(self, freq):
+        dataset = DatasetFactory(frequency=freq)
         assert_equal_dates(dataset.next_update,
                            datetime.now() + timedelta(days=7))
+
+    def test_next_update_biweekly(self):
+        dataset = DatasetFactory(frequency='biweekly')
+        assert_equal_dates(dataset.next_update,
+                           datetime.now() + timedelta(weeks=2))
+
+    def test_next_update_quarterly(self):
+        dataset = DatasetFactory(frequency='quarterly')
+        assert_equal_dates(dataset.next_update,
+                           datetime.now() + timedelta(days=365/4))
+
+    @pytest.mark.parametrize('freq', ['threeTimesAYear', 'semiannual', 'annual'])
+    def test_next_update_annual(self, freq):
+        dataset = DatasetFactory(frequency=freq)
+        assert_equal_dates(dataset.next_update,
+                           datetime.now() + timedelta(days=365))
+
+    def test_next_update_biennial(self):
+        dataset = DatasetFactory(frequency='biennial')
+        assert_equal_dates(dataset.next_update,
+                           datetime.now() + timedelta(days=365*2))
+
+    def test_next_update_triennial(self):
+        dataset = DatasetFactory(frequency='triennial')
+        assert_equal_dates(dataset.next_update,
+                           datetime.now() + timedelta(days=365*3))
+
+    def test_next_update_quinquennial(self):
+        dataset = DatasetFactory(frequency='quinquennial')
+        assert_equal_dates(dataset.next_update,
+                           datetime.now() + timedelta(days=365*5))
+
+    @pytest.mark.parametrize('freq', ['continuous', 'punctual', 'irregular', 'unknown'])
+    def test_next_update_undefined(self, freq):
+        dataset = DatasetFactory(frequency=freq)
+        assert dataset.next_update is None
 
     def test_quality_default(self):
         dataset = DatasetFactory(description='')
@@ -142,10 +191,15 @@ class DatasetModelTest:
             'score': 0
         }
 
-    def test_quality_next_update(self):
-        dataset = DatasetFactory(description='', frequency='weekly')
-        assert dataset.quality['update_fulfilled_in_time'] is True
+    @pytest.mark.parametrize('freq', UPDATE_FREQUENCIES)
+    def test_quality_frequency_update(self, freq):
+        dataset = DatasetFactory(description='', frequency=freq)
+        if freq == 'unknown':
+            assert dataset.quality['update_frequency'] is False
+            assert 'update_fulfilled_in_time' not in dataset.quality
+            return
         assert dataset.quality['update_frequency'] is True
+        assert dataset.quality['update_fulfilled_in_time'] is True
         assert dataset.quality['score'] == Dataset.normalize_score(2)
 
     def test_quality_description_length(self):
@@ -521,3 +575,66 @@ class ResourceSchemaTest:
         rmock.get('https://example.com/schemas', status_code=500)
         assert 'dummy_from_cache' == ResourceSchema.objects()
         assert rmock.call_count == 2
+
+
+class HarvestMetadataTest:
+    def test_harvest_dataset_metadata_validate_success(self):
+        dataset = DatasetFactory()
+
+        harvest_metadata = HarvestDatasetMetadata(
+            backend='DCAT',
+            created_at=datetime.now(),
+            modified_at=datetime.now(),
+            source_id='source_id',
+            remote_id='remote_id',
+            domain='domain.gouv.fr',
+            last_update=datetime.now(),
+            remote_url='http://domain.gouv.fr/dataset/remote_url',
+            uri='http://domain.gouv.fr/dataset/uri',
+            dct_identifier='http://domain.gouv.fr/dataset/identifier',
+            archived_at=datetime.now(),
+            archived='not-on-remote'
+        )
+        dataset.harvest = harvest_metadata
+        dataset.save()
+
+    def test_harvest_dataset_metadata_validation_error(self):
+        harvest_metadata = HarvestDatasetMetadata(created_at='maintenant')
+        dataset = DatasetFactory()
+        dataset.harvest = harvest_metadata
+        with pytest.raises(db.ValidationError):
+            dataset.save()
+
+    def test_harvest_dataset_metadata_no_validation_dynamic(self):
+        # Adding a dynamic field (not defined in HarvestDatasetMetadata) does not raise error
+        # at validation time
+        harvest_metadata = HarvestDatasetMetadata(dynamic_created_at='maintenant')
+        dataset = DatasetFactory()
+        dataset.harvest = harvest_metadata
+        dataset.save()
+
+    def test_harvest_resource_metadata_validate_success(self):
+        resource = ResourceFactory()
+
+        harvest_metadata = HarvestResourceMetadata(
+            created_at=datetime.now(),
+            modified_at=datetime.now(),
+            uri='http://domain.gouv.fr/dataset/uri'
+        )
+        resource.harvest = harvest_metadata
+        resource.validate()
+
+    def test_harvest_resource_metadata_validation_error(self):
+        harvest_metadata = HarvestResourceMetadata(created_at='maintenant')
+        resource = ResourceFactory()
+        resource.harvest = harvest_metadata
+        with pytest.raises(db.ValidationError):
+            resource.validate()
+
+    def test_harvest_resource_metadata_no_validation_dynamic(self):
+        # Adding a dynamic field (not defined in HarvestResourceMetadata) does not raise error
+        # at validation time
+        harvest_metadata = HarvestResourceMetadata(dynamic_created_at='maintenant')
+        resource = ResourceFactory()
+        resource.harvest = harvest_metadata
+        resource.validate()